pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "computer-use"
version = "0.1.0"
description = "AI-powered GUI automation tool"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyautogui>=0.9.54",
    "Pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "termcolor>=2.4.0",
]

[project.scripts]
computer-use = "computer_use.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.hatch.build.targets.wheel]
packages = ["src/computer_use"]
